backend/requirements.txt
fastapi
uvicorn
appwrite
pydantic
pydantic-settings
firebase-admin
bcrypt
PyJWT
